requirements.txt
Django==1.6.7
Mezzanine==3.1.10
MySQL-python==1.2.5
Paste==1.7.5.1
Pillow==2.5.3
South==1.0
argparse==1.2.1
beautifulsoup4==4.1.3
bleach==1.4
certifi==14.05.14
django-annoying==0.8.0
django-embed-video==0.11
django-mptt==0.6.1
django-ordered-model==0.4.0
git+git://github.com/renyi/django-pageviews.git
filebrowser-safe==0.3.6
future==0.9.0
gdata==2.0.18
grappelli-safe==0.3.13
html5lib==1.0b3
oauthlib==0.6.3
python-openid==2.2.5
python-social-auth==0.1.26
pytz==2014.7
requests==2.4.0
requests-oauthlib==0.4.1
six==1.7.3
tzlocal==1.0
wsgiref==0.1.2
